feat: refactor commands structure, add role notes
