+++ commands.ts
@@ -1,4 +1,4 @@
-import * as aftermath from './aftermath/index';
+import * as aftermath from './commands/aftermath/index';
 import * as faq from './commands/faq';
 import * as games from './commands/roles';
