+++ commands/aftermath/index.ts
