+++ commands/roles/notes.ts
@@ -1,0 +1,16 @@
+import { MessageEmbed } from 'discord.js';
+
+export const items = [
+  {
+    keyword: 'arma',
+    embed: new MessageEmbed()
+      .setTitle(`Arma 3 Ukraine`)
+      .setDescription(
+        'Також, запрошуємо завітати на діскорд сервер спільноти, яка створена для зібрання воєдино всіх українських гравців Arma 3.\n\n*"Ми цінуємо свободу та єдність."*\n\nhttps://discord.gg/DYA3dNS',
+      )
+      .setURL('https://discord.gg/DYA3dNS')
+      .setThumbnail(
+        'https://cdn.discordapp.com/icons/381716255431720980/9ef62297cf55d5d37df9fbbc40a7c1cf.webp',
+      ),
+  },
+];
